update example name and tidy up seed-data.tsx
