+++ examples/custom-id/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@keystone-6/example-custom-id",
-  "version": "0.0.6",
+  "version": "0.0.1",
   "private": true,
   "license": "MIT",
   "scripts": {
@@ -19,6 +19,5 @@
     "prisma": "^4.15.0",
     "tsx": "^3.9.0",
     "typescript": "~5.0.0"
-  },
-  "repository": "https://github.com/keystonejs/keystone/tree/main/examples/usecase-todo"
+  }
 }
