+++ examples/custom-id/schema.graphql
