+++ examples/custom-id/schema.ts
